xstage_books/requirements.txt
cryptography==43.0.0
Django==5.0.7
django-allauth==64.0.0
djangorestframework==3.15.2
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3